pyproject.toml
[project]
name = "wiki-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "mcp>=1.22.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "wikipedia-mcp>=1.6.0",
]
